Safari-609.1.11.1/LayoutTests/css3/flexbox/flex-one-sets-flex-basis-to-zero-px.html

<!DOCTYPE html>
<html>
<link href="resources/flexbox.css" rel="stylesheet">
<style>
.flexbox > div {
    font-family: Ahem;
    font-size: 14px;
}
    
.flex-one-one-zero-percent {
    flex: 1 1 0%;
    min-width: 0;
    min-height: 0;
}

.flex-one-one-zero-px {
    flex: 1 1 0px;
    min-width: 0;
    min-height: 0;
}

.vertical {
    writing-mode: vertical-rl;
}
</style>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<script src="../../resources/ahem.js"></script>
<body onload="checkLayout('.flexbox');">
<div id=log></div>
<div class="flexbox column">
  <div class="flex-one" data-expected-height="14">Flex item with flex: 1</div>
</div>

<div class="flexbox column">
  <div class="flex-one-one-zero-percent" data-expected-height="14">Flex item with flex: 1 1 0%</div>
</div>

<div class="flexbox column">
  <div class="flex-one-one-zero-px" data-expected-height="0">Flex item with flex: 1 1 0px</div>
</div>

<div class="flexbox column vertical">
  <div class="flex-one" data-expected-width="14">Flex item with flex: 1</div>
</div>

<div class="flexbox column vertical">
  <div class="flex-one-one-zero-percent" data-expected-width="14">Flex item with flex: 1 1 0%</div>
</div>

<div class="flexbox column vertical">
  <div class="flex-one-one-zero-px" data-expected-width="0">Flex item with flex: 1 1 0px</div>
</div>
</body>
</html>
